+++ components/pryv/src/index.js
@@ -9,7 +9,8 @@
  * @property {pryv.Connection} Connection - To interact with an individual's (user) data set
  * @property {pryv.Browser} Browser - Browser Tools - Access request helpers and visuals (button)
  * @property {pryv.utils} utils - Exposes some utils for HTTP calls and tools to manipulate Pryv's API endpoints
- * @property {pryv.PryvError} PryvError - Custom error class with innerObject support
+ * @property {pryv.PryvError} PryvError - Custom error class with innerObject + structured API-error fields
+ * @property {Object} ERRORS - Catalogue of Pryv API error ids (mirrors open-pryv.io/components/errors)
  */
 module.exports = {
   Service: require('./Service'),
@@ -18,5 +19,6 @@ module.exports = {
   Browser: require('./Browser'),
   utils: require('./utils'),
   PryvError: require('./lib/PryvError'),
+  ERRORS: require('./lib/errorIds'),
   version: require('../package.json').version
 };